# openapi.yaml
openapi: "3.0.3"
info:
  title: Bridage API
  version: "1.0.0"
  description: |
    **Bridage** is a server-side API-key relay proxy. It wraps upstream LLM provider
    API keys, generates downstream `brg_`-prefixed keys, and exposes an
    OpenAI-compatible `/v1` surface.

    ## Authentication

    | Scope | Mechanism |
    |-------|-----------|
    | Admin endpoints (`/admin/*`) | JWT Bearer token obtained from `POST /admin/login` |
    | Relay endpoints (`/v1/*`) | `brg_`-prefixed downstream key as Bearer token |
servers:
  - url: "http://localhost:8080"
    description: Local development

# ─── Security schemes ─────────────────────────────────────────────────────────
components:
  securitySchemes:
    AdminJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT issued by POST /admin/login. Pass as `Authorization: Bearer <token>`."

    BridageKey:
      type: http
      scheme: bearer
      description: "Downstream `brg_` key. Pass as `Authorization: Bearer brg_...`."

  # ─── Reusable schemas ────────────────────────────────────────────────────
  schemas:
...
    Model:
      type: object
      properties:
        id:              { type: string, format: uuid }
        provider_id:     { type: string, format: uuid }
        name:            { type: string, example: "gpt-4o" }
        provider_model:  { type: string, example: "gpt-4o-2024-11-20" }
        description:     { type: string }
        enabled:         { type: boolean }
        supports_images: { type: boolean }
        context_window:  { type: integer, example: 128000 }
        created_at:      { type: string, format: date-time }
        updated_at:      { type: string, format: date-time }
    ModelCreate:
      type: object
      required: [provider_id, name, provider_model]
      properties:
        provider_id:     { type: string, format: uuid }
        name:            { type: string, example: "gpt-4o" }
        provider_model:  { type: string, example: "gpt-4o-2024-11-20" }
        description:     { type: string }
        enabled:         { type: boolean, default: true }
        supports_images: { type: boolean, default: false }
        context_window:  { type: integer, example: 128000 }
...
    UsageLog:
      type: object
      properties:
        id:                 { type: string, format: uuid }
        api_key_id:         { type: string, format: uuid }
        endpoint:           { type: string, example: "/v1/chat/completions" }
        model:              { type: string, example: "gpt-4o" }
        prompt_tokens:      { type: integer }
        completion_tokens:  { type: integer }
        total_tokens:       { type: integer }
        status_code:        { type: integer }
        duration_ms:        { type: integer }
        error:              { type: string, nullable: true }
        created_at:         { type: string, format: date-time }
    UsageLogList:
      type: object
      properties:
        object: { type: string, example: "list" }
        data:
          type: array
          items: { $ref: "#/components/schemas/UsageLog" }

    # ── Preset ────────────────────────────────────────────────────────────
...
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          oneOf:
            - type: string
            - type: array
              items: { type: object }
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:       { type: string, example: "gpt-4o" }
        messages:
          type: array
          items: { $ref: "#/components/schemas/ChatMessage" }
        stream:      { type: boolean, default: false }
        temperature: { type: number, minimum: 0, maximum: 2 }
        max_tokens:  { type: integer }
        top_p:       { type: number }
        n:           { type: integer }
    EmbeddingsRequest:
      type: object
      required: [model, input]
      properties:
        model: { type: string }
        input:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
...
paths:
...
  /admin/models:
    get:
      summary: List all models
      operationId: listModels
      tags: [Admin / Models]
      security: [{ AdminJWT: [] }]
      responses:
        "200":
          description: Array of models.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Model" }
        "401": { description: Unauthorized. }
    post:
      summary: Create a model
      operationId: createModel
      tags: [Admin / Models]
      security: [{ AdminJWT: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModelCreate" }
      responses:
        "201":
          description: Model created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Model" }
        "400": { description: Validation error. }
        "401": { description: Unauthorized. }
...
  /v1/chat/completions:
    post:
      summary: Chat completions (OpenAI-compatible)
      operationId: chatCompletions
      tags: [Relay / Chat]
      security: [{ BridageKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChatCompletionRequest" }
      responses:
        "200":
          description: "Chat completion response (or SSE stream when `stream: true`)."
          content:
            application/json:
              schema: { type: object }
            text/event-stream:
              schema: { type: string }
        "401": { description: Invalid or missing brg_ key. }
        "429": { description: Rate limit or quota exceeded. }
        "502": { description: Upstream provider error. }
  /v1/responses:
    post:
      summary: Responses API (OpenAI Responses format)
      operationId: responses
      tags: [Relay / Chat]
      security: [{ BridageKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: Response object from upstream. }
        "401": { description: Unauthorized. }
        "429": { description: Rate limit or quota exceeded. }
  /v1/embeddings:
    post:
      summary: Embeddings (OpenAI-compatible)
      operationId: embeddings
      tags: [Relay / Embeddings]
      security: [{ BridageKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmbeddingsRequest" }
      responses:
        "200": { description: Embeddings response from upstream. }
        "401": { description: Unauthorized. }
        "429": { description: Rate limit or quota exceeded. }
...
  /v1/account/key:
    get:
      summary: Get my key's metadata
      description: Returns the metadata of the `brg_` key used for this request.
      operationId: getAccountKey
      tags: [Relay / Account]
      security: [{ BridageKey: [] }]
      responses:
        "200":
          description: Key metadata.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:              { type: string, format: uuid }
                  name:            { type: string }
                  status:          { type: string }
                  expires_at:      { type: string, format: date-time, nullable: true }
                  rate_limit:      { type: integer }
                  quota_tokens:    { type: integer }
                  used_tokens:     { type: integer }
                  quota_requests:  { type: integer }
                  used_requests:   { type: integer }
                  provider_id:     { type: string, format: uuid, nullable: true }
                  created_at:      { type: string, format: date-time }
        "401": { description: Unauthorized. }
  /v1/account/usage:
    get:
      summary: Get my recent usage history
      description: Returns the last 50 usage log entries for the authenticated key.
      operationId: getAccountUsage
      tags: [Relay / Account]
      security: [{ BridageKey: [] }]
      responses:
        "200":
          description: Usage log list.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsageLogList" }
        "401": { description: Unauthorized. }
...